mostly fleshed out how the multimodal model object is supposed to work, only thing left to do is actually try the code
